Change the dependency on HPPC to be 0.7.3.
This means renaming all uses of *OpenHash{Map,Set} to *Hash{Map,Set}.

This is forked from the 1.3.1 release, and the version number has been
set to 1.3.99-SNAPSHOT, to make it clear that this is not an official
release.

This gives me something that is functionally equivalent to 1.3.1, but
with the different HPPC dependency.  This is something that I can use with
the cTakes 4.0.0 release to move to HPPC 0.7.3 without making major changes
to cTakes.